LayoutTests/http/tests/dom/document-attributes-null-handling.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
shouldBeEqualToString("document.cookie", "");
evalAndLog("document.cookie = 'key=value'");
shouldBeEqualToString("document.cookie", "key=value");
evalAndLog("document.cookie = 'foo'");
shouldBeTrue("document.cookie.indexOf('key=value') != -1");
shouldBeTrue("document.cookie.indexOf('null=') != -1");
debug(document.cookie)

debug("");
debug("* title attribute");
shouldBeEqualToString("document.title", "");
evalAndLog("document.title = null");
shouldBeEqualToString("document.title", "null");

debug("");
debug("* domain attribute");
shouldBeEqualToString("document.domain", "127.0.0.1");
shouldThrow("document.domain = null");
shouldBeEqualToString("document.domain", "127.0.0.1");
</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
